Removed Obfuscator class and moved logic to proccessor
